Refactor actions into folders
