+++ activity_browser/actions/activity/activity_duplicate.py
@@ -2,9 +2,9 @@ from typing import Union, Callable, List
 
 from PySide2 import QtCore
 
-from ..controllers.activity import activity_controller
-from ..ui.icons import qicons
-from .base import ABAction
+from activity_browser.controllers.activity import activity_controller
+from activity_browser.ui.icons import qicons
+from activity_browser.actions.base import ABAction
 
 
 class ActivityDuplicate(ABAction):
