+++ activity_browser/actions/activity/activity_open.py
@@ -3,8 +3,8 @@ from typing import Union, Callable, List
 from PySide2 import QtCore
 
 from activity_browser import signals
-from ..ui.icons import qicons
-from .base import ABAction
+from activity_browser.ui.icons import qicons
+from activity_browser.actions.base import ABAction
 
 
 class ActivityOpen(ABAction):
