+++ activity_browser/actions/activity/activity_relink.py
@@ -5,9 +5,9 @@ from PySide2 import QtWidgets, QtCore
 
 from activity_browser import signals, application
 from activity_browser.bwutils.strategies import relink_activity_exchanges
-from .base import ABAction
-from ..ui.widgets import ActivityLinkingDialog, ActivityLinkingResultsDialog
-from ..ui.icons import qicons
+from activity_browser.actions.base import ABAction
+from activity_browser.ui.widgets import ActivityLinkingDialog, ActivityLinkingResultsDialog
+from activity_browser.ui.icons import qicons
 
 
 class ActivityRelink(ABAction):
