+++ activity_browser/actions/calculation_setup/cs_new.py
@@ -4,9 +4,9 @@ import brightway2 as bw
 from PySide2 import QtCore, QtWidgets
 
 from activity_browser import application
-from .base import ABAction
-from ..ui.icons import qicons
-from ..controllers import calculation_setup_controller
+from activity_browser.actions.base import ABAction
+from activity_browser.ui.icons import qicons
+from activity_browser.controllers import calculation_setup_controller
 
 
 class CSNew(ABAction):
