+++ activity_browser/actions/database/database_duplicate.py
@@ -4,10 +4,10 @@ import brightway2 as bw
 from PySide2 import QtWidgets, QtCore
 
 from activity_browser import application
-from .base import ABAction
-from ..ui.icons import qicons
-from ..ui.threading import ABThread
-from ..controllers import database_controller
+from activity_browser.actions.base import ABAction
+from activity_browser.ui.icons import qicons
+from activity_browser.ui.threading import ABThread
+from activity_browser.controllers import database_controller
 
 
 class DatabaseDuplicate(ABAction):
